Refactor elmundo parser
elmundo parser module 2.0. First finished and polished parser module.
It parses all the newspaper and it cleans unnecessary stuff.
To get the links of the articles that might be parsed it's used regex.
This is due to the fast search, easy to use and package independence it provides.
To clean all the mess it's used a npm package called node-readability.
Using this package lets us be more lazy in unecessary stuff as it gets rid of
getting the Html source of every article, and it cleans most of the ads.
What is not cleaned by the package has been cleaned later by using traditional
methods as .split

 much work, as getting every source page and