Refactor news-parser/index.js http https implement and tmp supress
Supresion of the parser package used untill now to get the webpage content
from the websites parsed. There is no need of external packages, fewer
dependencies => lower maintanance, and we get the main html faster than ever
We start using http and https built in libs

Refactor of the tmp files, untill now we were using tmp files, but as
we want to send files attached in emails we need to get rid of tmp files.
For now I have implemented a hashid.epub|.mobi code. In the future it should
be something like website_parsed.epub & .mobi