Add sending-email with attachment, Kindle autosynch
Must have feature for kindle owners who will now recieve their
bundles with a single api call.

Add new helper function which is simple send-email function with
a single param(file_path). It's built in promises structure so that we can
have full control of it. If it had been implemented like an async function
it would have had problems with file deletion. This way it only removes files
when the files have been sent.

Refactor of the user endpoint so that it can send this new-kind of messages
and later delete those bundle files.

Later implementation will delete the bundle once it generetes the
new bundle. And it will also send this email automatically without
having to call the api endpoint.