Make Heroku Api implementation
Since AWS ec2 docker service was a bit tricky so time to move on.
Heroku is a nice service to use as we don't want to spend too much time
on devops while still working on the basic product. Free tier avialable.
Just push the project to heroku and it works.